fix up exports to be easier to use
