+++ crates/coprocessor-node/src/job_processor.rs
@@ -3,11 +3,11 @@
 use crate::{metrics::Metrics, relayer::JobRelayer};
 use alloy::{hex, primitives::Signature, signers::Signer};
 use async_channel::Receiver;
+use ivm_abi::abi_encode_offchain_job_request;
 use ivm_db::{
     delete_fail_relay_job, get_all_failed_jobs, put_fail_relay_job, put_job,
     tables::{ElfWithMeta, Job, RequestType},
 };
-use ivm_abi::abi_encode_offchain_job_request;
 use ivm_proto::{JobStatus, JobStatusType, VmType};
 use reth_db::Database;
 use std::{marker::Send, sync::Arc, time::Duration};
