+++ crates/coprocessor-node/src/server.rs
@@ -7,8 +7,8 @@ use alloy::{
     signers::Signer,
     sol_types::SolType,
 };
-use ivm_db::tables::{Job, RequestType};
 use ivm_abi::{get_job_id, OffchainJobRequest};
+use ivm_db::tables::{Job, RequestType};
 use ivm_proto::{
     coprocessor_node_server::CoprocessorNode as CoprocessorNodeTrait, GetResultRequest,
     GetResultResponse, JobResult, JobStatus, JobStatusType, SubmitJobRequest, SubmitJobResponse,
