+++ crates/mock-consumer/src/lib.rs
@@ -8,11 +8,10 @@ use alloy::{
     sol_types::SolValue,
 };
 use contracts::mock_consumer::MockConsumer;
+use ivm_abi::{abi_encode_result_with_metadata, get_job_id};
 use ivm_db::tables::{Job, RequestType};
-use ivm_abi::get_job_id;
 use ivm_proto::{JobStatus, JobStatusType};
 use ivm_test_utils::{get_signers, AnvilJobManager};
-use zkvm_executor::service::abi_encode_result_with_metadata;
 
 /// Max cycles that the `MockContract` calls create job with.
 pub const MOCK_CONSUMER_MAX_CYCLES: u64 = 1_000_000;
