+++ crates/scripts/src/signer.rs
@@ -6,13 +6,12 @@ use alloy::{
 };
 use dotenv::dotenv;
 use eip4844::{SidecarBuilder, SimpleCoder};
-use ivm_abi::get_job_id;
+use ivm_abi::{
+    abi_encode_offchain_result_with_metadata, abi_encode_result_with_metadata, get_job_id,
+};
 use ivm_test_utils::create_and_sign_offchain_request;
 use k256::ecdsa::SigningKey;
 use std::env;
-use zkvm_executor::service::{
-    abi_encode_offchain_result_with_metadata, abi_encode_result_with_metadata,
-};
 
 type K256LocalSigner = LocalSigner<SigningKey>;
