+++ crates/sdk/abi/src/lib.rs
@@ -1,7 +1,7 @@
 //! Common crate for ABI-encoded types.
 
 use alloy::{
-    primitives::{keccak256, Address},
+    primitives::{keccak256, Address, FixedBytes},
     sol,
     sol_types::{SolType, SolValue},
 };
@@ -47,6 +47,42 @@ sol! {
         bytes32 output_state_root;
         bytes result;
     }
+
+    /// The payload that gets signed to signify that the zkvm executor has faithfully
+    /// executed the job. Also the result payload the job manager contract expects.
+    #[derive(Default, PartialEq, Eq, PartialOrd, Ord, Debug)]
+    struct ResultWithMetadata {
+        /// Job ID.
+        bytes32 job_id;
+        /// Hash of onchain input passed to zkVM program for this job.
+        bytes32 onchain_input_hash;
+        /// Max cycles for the job.
+        uint64 max_cycles;
+        /// Program ID of program being executed.
+        bytes program_id;
+        /// Raw output (result) from zkVM program.
+        bytes raw_output;
+    }
+
+    /// The payload that gets signed to signify that the zkvm executor has faithfully
+    /// executed the offchian job. Also the result payload the job manager contract expects.
+    #[derive(Default, PartialEq, Eq, PartialOrd, Ord, Debug)]
+    struct OffchainResultWithMetadata {
+        /// Job ID.
+        bytes32 job_id;
+        /// Hash of onchain input passed to zkVM program for this job.
+        bytes32 onchain_input_hash;
+        /// Hash of offchain input passed to zkVM program for this job.
+        bytes32 offchain_input_hash;
+        /// Max cycles for the job.
+        uint64 max_cycles;
+        /// Program ID of program being executed.
+        bytes program_id;
+        /// Raw output (result) from zkVM program.
+        bytes raw_output;
+        /// Commitments to the blobs that will contain the offchain input.
+        bytes32[] versioned_blob_hashes;
+    }
 }
 
 /// Returns an ABI-encoded offchain job request. This ABI-encoded response will be
@@ -76,4 +112,46 @@ pub type NonceAndConsumer = sol! {
 /// Convenience method to abi encode [`NonceAndConsumer`].
 pub fn abi_encode_nonce_and_consumer(nonce: u64, consumer: Address) -> Vec<u8> {
     NonceAndConsumer::abi_encode_packed(&(nonce, consumer))
+}
+
+/// Returns an ABI-encoded result with metadata. This ABI-encoded response will be
+/// signed by the operator.
+pub fn abi_encode_result_with_metadata(
+    job_id: [u8; 32],
+    onchain_input_hash: FixedBytes<32>,
+    max_cycles: u64,
+    program_id: &[u8],
+    raw_output: &[u8],
+) -> Vec<u8> {
+    ResultWithMetadata {
+        job_id: job_id.into(),
+        onchain_input_hash,
+        max_cycles,
+        program_id: program_id.to_vec().into(),
+        raw_output: raw_output.to_vec().into(),
+    }
+    .abi_encode()
+}
+
+/// Returns an ABI-encoded offchain result with metadata. This ABI-encoded response will be
+/// signed by the operator.
+pub fn abi_encode_offchain_result_with_metadata(
+    job_id: [u8; 32],
+    onchain_input_hash: FixedBytes<32>,
+    offchain_input_hash: FixedBytes<32>,
+    max_cycles: u64,
+    program_id: &[u8],
+    raw_output: &[u8],
+    versioned_blob_hashes: Vec<FixedBytes<32>>,
+) -> Vec<u8> {
+    OffchainResultWithMetadata {
+        job_id: job_id.into(),
+        onchain_input_hash,
+        offchain_input_hash,
+        max_cycles,
+        program_id: program_id.to_vec().into(),
+        raw_output: raw_output.to_vec().into(),
+        versioned_blob_hashes,
+    }
+    .abi_encode()
 }
